requirements.txt
pygame==2.6.1
requests==2.32.3
pyaudio==0.2.14
webrtcvad==2.0.10
numpy==1.22.0
torch==2.6.0
google-generativeai==0.8.4
git+https://github.com/openai/whisper.git
